Fixing type-mismatch errors in power10 sandbox
There was a mismatch between the definition and declaration of
bli_gemm_ex() function in the power10 sandbox, which was
throwing compilation errors. Now, the definition matches the declaration
in blis.h.

There was also a mismatch between the function definition and the
function call for bli_gemm_front() in power10 sandbox. This is also
fixed by typecasting the corresponding argument passed in the
function.